удалил комменты и placeholder; создал отдельный файл для PaddedTextField
